Removed some cookie cutter parameters: - `use_heroku`, - `keep_local_envs_in_vcs`, - `windows`, - `use_pycharm`, - `use_docker`, - `postgresql_version`, - `js_task_runner`, - `custom_bootstrsp_compilation`, - `debug`
