+++ cookiecutter.json
@@ -2,7 +2,7 @@
   "project_name": "My Awesome Project",
   "project_slug": "{{ cookiecutter.project_name.lower()|replace(' ', '_')|replace('-', '_')|replace('.', '_')|trim() }}",
   "description": "Behold My Awesome Project!",
-  "author_name": "Daniel Roy Greenfeld",
+  "author_name": "Your Name",
   "domain_name": "example.com",
   "email": "{{ cookiecutter.author_name.lower()|replace(' ', '-') }}@example.com",
   "version": "0.1.0",
@@ -14,34 +14,16 @@
     "Not open source"
   ],
   "timezone": "UTC",
-  "windows": "n",
-  "use_pycharm": "n",
-  "use_docker": "n",
-  "postgresql_version": [
-    "11.3",
-    "10.8",
-    "9.6",
-    "9.5",
-    "9.4"
-  ],
-  "js_task_runner": [
-    "None",
-    "Gulp"
-  ],
   "cloud_provider": [
     "AWS",
+    "Azure",
     "GCP",
     "None"
   ],
-  "custom_bootstrap_compilation": "n",
   "use_compressor": "n",
   "use_celery": "n",
   "use_mailhog": "n",
   "use_sentry": "n",
   "use_whitenoise": "n",
-  "use_heroku": "n",
-  "use_travisci": "n",
-  "keep_local_envs_in_vcs": "y",
-
-  "debug": "n"
+  "use_travisci": "n"
 }
