+++ docs/project-generation-options.rst
@@ -37,29 +37,6 @@ open_source_license:
 timezone:
     The value to be used for the ``TIME_ZONE`` setting of the project.
 
-windows:
-    Indicates whether the project should be configured for development on Windows.
-
-use_pycharm:
-    Indicates whether the project should be configured for development with PyCharm_.
-
-use_docker:
-    Indicates whether the project should be configured to use Docker_ and `Docker Compose`_.
-
-postgresql_version:
-    Select a PostgreSQL_ version to use. The choices are:
-
-    1. 11.3
-    2. 10.8
-    3. 9.6
-    4. 9.5
-    5. 9.4
-
-js_task_runner:
-    Select a JavaScript task runner. The choices are:
-
-    1. None
-    2. Gulp_
 
 cloud_provider:
     Select a cloud provider for static & media files. The choices are:
@@ -70,10 +47,6 @@ cloud_provider:
 
     Note that if you choose no cloud provider, media files won't work.
 
-custom_bootstrap_compilation:
-    Indicates whether the project should support Bootstrap recompilation
-    via the selected JavaScript task runner's task. This can be useful
-    for real-time Bootstrap variable alteration.
 
 use_compressor:
     Indicates whether the project should be configured to use `Django Compressor`_.
@@ -96,17 +69,6 @@ use_heroku:
 
 use_travisci:
     Indicates whether the project should be configured to use `Travis CI`_.
-
-keep_local_envs_in_vcs:
-    Indicates whether the project's ``.envs/.local/`` should be kept in VCS
-    (comes in handy when working in teams where local environment reproducibility
-    is strongly encouraged).
-    Note: .env(s) are only utilized when Docker Compose and/or Heroku support is enabled.
-
-debug:
-    Indicates whether the project should be configured for debugging.
-    This option is relevant for Cookiecutter Django developers only.
-
 
 .. _MIT: https://opensource.org/licenses/MIT
 .. _BSD: https://opensource.org/licenses/BSD-3-Clause
