+++ tox.ini
@@ -16,4 +16,4 @@ commands = pytest -m black {posargs:./tests}
 
 [testenv:black-template]
 deps = black
-commands = black --check hooks tests setup.py docs
+commands = black --diff --check hooks tests setup.py docs
