+++ {{cookiecutter.project_slug}}/.gitignore
@@ -161,7 +161,6 @@ typings/
 !.vscode/extensions.json
 
 
-{% if cookiecutter.use_pycharm == 'y' -%}
 # Provided default Pycharm Run/Debug Configurations should be tracked by git
 # In case of local modifications made by Pycharm, use update-index command
 # for each changed file, like this:
@@ -216,7 +215,6 @@ com_crashlytics_export_strings.xml
 crashlytics.properties
 crashlytics-build.properties
 fabric.properties
-{% endif %}
 
 
 ### Windows template
@@ -321,23 +319,19 @@ Session.vim
 
 # Auto-generated tag files
 tags
-{% if cookiecutter.use_docker == 'n' %}
 
 ### VirtualEnv template
 # Virtualenv
 pyvenv.cfg
 pip-selfcheck.json
 .env
-{% endif %}
 
 ### Project template
-{% if cookiecutter.use_mailhog == 'y' and cookiecutter.use_docker == 'n' %}
+{% if cookiecutter.use_mailhog == 'y' %}
 MailHog
 {%- endif %}
 {{ cookiecutter.project_slug }}/media/
 
 .pytest_cache/
 
-{% if cookiecutter.use_docker == 'y' %}
 .ipython/
-{%- endif %}
