+++ {{cookiecutter.project_slug}}/.idea/{{cookiecutter.project_slug}}.iml
@@ -13,13 +13,7 @@
     </facet>
   </component>
   <component name="NewModuleRootManager">
-    {% if cookiecutter.js_task_runner != 'None' %}
-    <content url="file://$MODULE_DIR$"> 
-      <excludeFolder url="file://$MODULE_DIR$/node_modules" />
-    </content>
-    {% else %} 
-    <content url="file://$MODULE_DIR$" /> 
-    {% endif %}
+    <content url="file://$MODULE_DIR$" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
   <component name="PackageRequirementsSettings">
