+++ {{cookiecutter.project_slug}}/gulpfile.js
@@ -9,9 +9,6 @@ const pjson = require('./package.json')
 // Plugins
 const autoprefixer = require('autoprefixer')
 const browserSync = require('browser-sync').create()
-{% if cookiecutter.custom_bootstrap_compilation == 'y' %}
-const concat = require('gulp-concat')
-{% endif %}
 const cssnano = require ('cssnano')
 const imagemin = require('gulp-imagemin')
 const pixrem = require('pixrem')
@@ -29,14 +26,6 @@ function pathsConfig(appName) {
   const vendorsRoot = 'node_modules'
 
   return {
-    {% if cookiecutter.custom_bootstrap_compilation == 'y' %}
-    bootstrapSass: `${vendorsRoot}/bootstrap/scss`,
-    vendorsJs: [
-      `${vendorsRoot}/jquery/dist/jquery.slim.js`,
-      `${vendorsRoot}/popper.js/dist/umd/popper.js`,
-      `${vendorsRoot}/bootstrap/dist/js/bootstrap.js`,
-    ],
-    {% endif %}
     app: this.app,
     templates: `${this.app}/templates`,
     css: `${this.app}/static/css`,
@@ -67,9 +56,6 @@ function styles() {
   return src(`${paths.sass}/project.scss`)
     .pipe(sass({
       includePaths: [
-        {% if cookiecutter.custom_bootstrap_compilation == 'y' %}
-        paths.bootstrapSass,
-        {% endif %}
         paths.sass
       ]
     }).on('error', sass.logError))
@@ -89,19 +75,6 @@ function scripts() {
     .pipe(rename({ suffix: '.min' }))
     .pipe(dest(paths.js))
 }
-
-{% if cookiecutter.custom_bootstrap_compilation == 'y' %}
-// Vendor Javascript minification
-function vendorScripts() {
-  return src(paths.vendorsJs)
-    .pipe(concat('vendors.js'))
-    .pipe(dest(paths.js))
-    .pipe(plumber()) // Checks for errors
-    .pipe(uglify()) // Minifies the js
-    .pipe(rename({ suffix: '.min' }))
-    .pipe(dest(paths.js))
-}
-{% endif %}
 
 // Image compression
 function imgCompression() {
@@ -128,22 +101,7 @@ function initBrowserSync() {
         `${paths.templates}/*.html`
       ], {
         // https://www.browsersync.io/docs/options/#option-proxy
-        {%- if cookiecutter.use_docker == 'n' %}
         proxy: 'localhost:8000'
-        {% else %}
-        proxy:  {
-          target: 'django:8000',
-          proxyReq: [
-            function(proxyReq, req) {
-              // Assign proxy "host" header same as current request at Browsersync server
-              proxyReq.setHeader('Host', req.headers.host)
-            }
-          ]
-        },
-        // https://www.browsersync.io/docs/options/#option-open
-        // Disable as it doesn't work from inside a container
-        open: false
-        {%- endif %}
       }
     )
 }
@@ -159,15 +117,12 @@ function watchPaths() {
 const generateAssets = parallel(
   styles,
   scripts,
-  {% if cookiecutter.custom_bootstrap_compilation == 'y' %}vendorScripts,{% endif %}
   imgCompression
 )
 
 // Set up dev environment
 const dev = parallel(
-  {%- if cookiecutter.use_docker == 'n' %}
   runServer,
-  {%- endif %}
   initBrowserSync,
   watchPaths
 )
