+++ {{cookiecutter.project_slug}}/pytest.ini
@@ -1,5 +1,3 @@
 [pytest]
 addopts = --ds=config.settings.test
-{%- if cookiecutter.js_task_runner != 'None' %}
 norecursedirs = node_modules
-{%- endif %}
