+++ {{cookiecutter.project_slug}}/requirements/base.txt
@@ -2,7 +2,7 @@ pytz==2019.1  # https://github.com/stub42/pytz
 python-slugify==3.0.2  # https://github.com/un33k/python-slugify
 Pillow==6.1.0  # https://github.com/python-pillow/Pillow
 {%- if cookiecutter.use_compressor == "y" %}
-rcssmin==1.0.6{% if cookiecutter.windows == 'y' and cookiecutter.use_docker == 'n' %} --install-option="--without-c-extensions"{% endif %}  # https://github.com/ndparker/rcssmin
+rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
 {%- endif %}
 argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
 {%- if cookiecutter.use_whitenoise == 'y' %}
@@ -12,9 +12,7 @@ redis==3.2.1  # https://github.com/antirez/redis
 {%- if cookiecutter.use_celery == "y" %}
 celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
 django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
-{%- if cookiecutter.use_docker == 'y' %}
 flower==0.9.3  # https://github.com/mher/flower
-{%- endif %}
 {%- endif %}
 
 # Django
