+++ {{cookiecutter.project_slug}}/requirements/local.txt
@@ -3,11 +3,7 @@
 Werkzeug==0.14.1  # pyup: < 0.15 # https://github.com/pallets/werkzeug
 ipdb==0.12  # https://github.com/gotcha/ipdb
 Sphinx==2.1.2  # https://github.com/sphinx-doc/sphinx
-{%- if cookiecutter.use_docker == 'y' %}
-psycopg2==2.8.3 --no-binary psycopg2  # https://github.com/psycopg/psycopg2
-{%- else %}
 psycopg2-binary==2.8.3  # https://github.com/psycopg/psycopg2
-{%- endif %}
 
 # Testing
 # ------------------------------------------------------------------------------
