+++ {{cookiecutter.project_slug}}/utility/install_python_dependencies.sh
@@ -35,7 +35,5 @@ if [ -z "$VIRTUAL_ENV" ]; then
 else
 
     pip install -r $PROJECT_DIR/requirements/local.txt
-    {% if cookiecutter.use_heroku == "y" -%}
     pip install -r $PROJECT_DIR/requirements.txt
-    {%- endif %}
 fi
