+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/celery_app.py
