+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/settings/__init__.py
@@ -1,0 +1,4 @@
+try:
+    from .dev import *  # NOQA
+except ImportError:
+    from .prod import *  # NOQA
