+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/settings/prod.py
@@ -1,0 +1,15 @@
+from .base import *  # noqa
+
+
+# TEMPLATES
+# https://docs.djangoproject.com/en/dev/ref/settings/#templates
+# ------------------------------------------------------------------------------
+TEMPLATES[0]["OPTIONS"]["loaders"] = [  # noqa F405
+    (
+        "django.template.loaders.cached.Loader",
+        [
+            "django.template.loaders.filesystem.Loader",
+            "django.template.loaders.app_directories.Loader",
+        ],
+    )
+]
