+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/settings/test.py
