+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/static/sass/project.scss
@@ -1,10 +1,3 @@
-{% if cookiecutter.custom_bootstrap_compilation == 'y' %}
-@import "custom_bootstrap_vars";
-@import "bootstrap";
-{% endif %}
-
-
-
 // project specific CSS goes here
 
 ////////////////////////////////
